renames invitee-poll-sections (component is shared between both poll views)
