+++ src/js/components/event/poll-sections.jsx
